% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enst2symbol.R
\name{enst2symbol}
\alias{enst2symbol}
\title{ENST to Symbol}
\usage{
enst2symbol(ENST)
}
\arguments{
\item{ENST}{input ENST ID}
}
\value{

}
\description{
ENST to Symbol
}
\examples{
enst2symbol("ENST00000228837")
}
